Move sample content and chunks to util
