+++ tests/test_file.py
@@ -4,24 +4,9 @@ from six.moves import builtins
 
 import resumable.file
 from resumable.file import LazyLoadChunkableFile
-
-
-SAMPLE_CONTENT = b'sample content afsdfas'
-TEST_CHUNK_SIZE = 10
-
-
-def expected_chunk_data():
-    start = 0
-    while start < len(SAMPLE_CONTENT):
-        yield SAMPLE_CONTENT[start:start+TEST_CHUNK_SIZE]
-        start += TEST_CHUNK_SIZE
-
-
-@pytest.fixture
-def sample_file(tmpdir):
-    path = tmpdir.join('sample-file.txt')
-    path.write(SAMPLE_CONTENT)
-    return path
+from tests.util import (
+    SAMPLE_CONTENT, TEST_CHUNK_SIZE, SAMPLE_CONTENT_CHUNKS, sample_file
+)
 
 
 @pytest.fixture
@@ -83,7 +68,7 @@ def test_read_bytes_lock(sample_file, mock_lock, mock_open):
 
 def test_chunking(sample_file):
     file = LazyLoadChunkableFile(sample_file, TEST_CHUNK_SIZE)
-    assert len(list(file.chunks)) == len(list(expected_chunk_data()))
-    for chunk, expected_data in zip(file.chunks, expected_chunk_data()):
+    assert len(list(file.chunks)) == len(SAMPLE_CONTENT_CHUNKS)
+    for chunk, expected_data in zip(file.chunks, SAMPLE_CONTENT_CHUNKS):
         assert chunk.data == expected_data
         assert chunk.size == len(expected_data)
